Add !alert and !role commands (#31)
* Add alert command

* Data migration

* Auto-close silently

* Add `!roles` command

* Whitespace fixes

* Consistency

* Add EOF newlines to files that are missing it

* Try to make this neater for Bsian

* Make inboxAdminRoleId checks consistent

* Return error if no message is found?

* Fix missing return

+++ src/modules/logs.js
@@ -57,14 +57,9 @@ module.exports = bot => {
       if (args[0] === "delete") {
         const userId = utils.getUserMention(args.slice(1).join(" "));
         if (! userId) return utils.postSystemMessageWithFallback(msg.channel, thread, "Please provide a user mention or ID!");
+        if (! msg.member.roles || ! msg.member.roles.includes(config.inboxAdminRoleId)) return;
 
-        if (! config.inboxAdminRoleId) {
-          return;
-        }
-
-        if (msg.member.roles && msg.member.roles.includes(config.inboxAdminRoleId)) {
-          deleteLogs(userId);
-        }
+        return deleteLogs(userId);
       }
 
       // User mention/id as argument
