+++ src/modules/purge.js
@@ -8,41 +8,38 @@ const threadUtils = require("../threadUtils");
 module.exports = bot => {
   threadUtils.addInboxServerCommand(bot, "purge", async (msg, args, thread) => {
     if (! thread) return;
+    if (! msg.member.roles || ! msg.member.roles.includes(config.inboxAdminRoleId)) return;
 
-    if (! config.inboxAdminRoleId) {
-		return;
+	// @ts-ignore
+	const channel = await thread.getDMChannel();
+	if (! channel || ! channel.id) {
+		return thread.postSystemMessage("Error getting DM Channel");
 	}
 
-	if (msg.member.roles && msg.member.roles.includes(config.inboxAdminRoleId)) {
-		// @ts-ignore
-		const channel = await thread.getDMChannel();
-		if (! channel || ! channel.id) {
-			return thread.postSystemMessage("Error getting DM Channel");
-		}
+	const msgs = await bot.getMessages(channel.id, 100);
+	let messages = msgs.filter(m => m.author.id === bot.user.id).map(m => m.id);
 
-		const msgs = await bot.getMessages(channel.id, 100);
-		let messages = msgs.filter(m => m.author.id === bot.user.id).map(m => m.id);
+	if (args.length > 0 && ! isNaN(Number(args[0]))) {
+		messages = messages.slice(0, parseInt(args[0], 10));
+	}
 
-		if (args.length > 0 && ! isNaN(Number(args[0]))) {
-			messages = messages.slice(0, parseInt(args[0], 10));
-		}
-
-		if (messages && messages.length) {
-			try {
-				const count = messages.length;
-				for (let id of messages) {
-					bot.deleteMessage(channel.id, id).catch(() => null);
-				}
-				thread.postSystemMessage(`Purged ${count} messages.`);
-			} catch (err) {
-				thread.postSystemMessage(`Error deleting messages: ${err.message}`);
+	if (messages && messages.length) {
+		try {
+			const count = messages.length;
+			for (let id of messages) {
+				bot.deleteMessage(channel.id, id).catch(() => null);
 			}
+			thread.postSystemMessage(`Purged ${count} messages.`);
+		} catch (err) {
+			thread.postSystemMessage(`Error deleting messages: ${err.message}`);
 		}
+	} else {
+		thread.postSystemMessage("I couldn't find any messages to delete.");
 	}
   });
 
 	bot.registerCommandAlias("p", "purge");
-	
+
 	threadUtils.addInboxServerCommand(bot, "undo", async (msg, args, thread) => {
 		if (! thread) return;
 
@@ -62,6 +59,8 @@ module.exports = bot => {
 			} catch (err) {
 				thread.postSystemMessage(`Error deleting messages: ${err.message}`);
 			}
+		} else {
+			thread.postSystemMessage("I couldn't find a message to delete.");
 		}
 	});
 };
